requirements.txt
# Data
pandas>=2.0
numpy>=1.24
yfinance>=0.2.30

# Technical Indicators
ta>=0.11

# Machine Learning
scikit-learn>=1.3
lightgbm>=4.0
shap>=0.43
optuna>=3.4

# Visualization & Dashboard
matplotlib>=3.7
seaborn>=0.13
plotly>=5.18
streamlit>=1.30

# Utilities
pyyaml>=6.0
joblib>=1.3
tqdm>=4.66
